Swap down to floats for faster speed

+++ benches/universe.rs
@@ -5,16 +5,16 @@ extern crate physics_sim;
 use physics_sim::*;
 use test::Bencher;
 
-const OBJECT_MASS: f64 = 291083.10129;
+const OBJECT_MASS: f32 = 291083.10129;
 const UNIVERSE_OBJECTS: usize = 10;
-const SPACING: f64 = 1000000.;
+const SPACING: f32 = 1000000.;
 
 #[bench]
 fn universe_creation(b: &mut Bencher) {
     b.iter(|| {
         let mut universe = Universe::default();
         for i in 0..UNIVERSE_OBJECTS {
-            universe.add_object(Object::new(OBJECT_MASS, Vector::default(), Point::new(i as f64 * SPACING, i as f64 * SPACING)));
+            universe.add_object(Object::new(OBJECT_MASS, Vector::default(), Point::new(i as f32 * SPACING, i as f32 * SPACING)));
         }
     });
 }
@@ -24,7 +24,7 @@ fn universe_update(b: &mut Bencher) {
     b.iter(|| {
         let mut universe = Universe::default();
         for i in 0..UNIVERSE_OBJECTS {
-            universe.add_object(Object::new(OBJECT_MASS, Vector::default(), Point::new(i as f64 * SPACING, i as f64 * SPACING)));
+            universe.add_object(Object::new(OBJECT_MASS, Vector::default(), Point::new(i as f32 * SPACING, i as f32 * SPACING)));
         }
         universe.update_state(1.);
     });
